Removed old Component. Documented IRLookup
